scripts/brightness/requirements.txt
tsl2561
RPi.GPIO==0.7.1
